+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##Next Version (Release Date TBC) Release Notes
 
+* **Issue #210** - When downloading transitive dependencies, optional dependencies are now not pulled down by default.  To change this toggle `<downloadOptionalDependencies>false</downloadOptionalDependencies>`
 * **Issue #179** - The plugin will now copy everything in the `${basedir}/src/test/conf` directory into the `jmeter/bin/` directory.  You can configure this by setting
 	```
 	<confFilesDirectory>${basedir}/src/test/some_conf_dir</confFilesDirectory>